Huge Database refactor
- Lots of changes on the database implementation which uses now a
generic interface for the libraries which allow the actual database
connector to be switched (only at compile time for now)
- Cascade changes to the rest of the codebase and CMakeLists
- Minor changes to the codebase regarding error reporting, initialization
of objects and generic coding style

+++ recon_daemon/Recon_settings.h
@@ -49,7 +49,7 @@ struct Recon_config{
     int request_chunk_size;
 
     /** if dry run is set recon daemon will only fetch certificates without importing them */
-    int dry_run;
+    unsigned long dry_run;
 
     /** flag used to ignore logging for known bugs */
     int ignore_known_bug;
